Pulled the data source and filter builders out into separate files
This simplifies the main file significantly.
