+++ source/components/cardContainer/container/index.ts
@@ -1,10 +1,7 @@
-import { CardSearchComponent } from './cardSearch/cardSearch';
-import { ItemCountComponent } from './itemCount/itemCount';
-import { SelectionComponent } from './selectionControl/selectionControl';
 import { ContainerHeaderTemplate } from './containerHeader.template';
 import { ContainerFooterTemplate } from './containerFooter.template';
 
-export const CONTAINER_DIRECTIVES: any[] = [CardSearchComponent, ItemCountComponent, SelectionComponent, ContainerHeaderTemplate, ContainerFooterTemplate];
+export const CONTAINER_DIRECTIVES: any[] = [ContainerHeaderTemplate, ContainerFooterTemplate];
 
 export * from './cardSearch/cardSearch';
 export * from './itemCount/itemCount';
