+++ source/components/cardContainer/index.ts
@@ -4,7 +4,7 @@ import * as filters from './filters/index';
 import * as paging from './paging/index';
 import * as sorts from './sorts/index';
 
-export const CARD_CONTAINER_DIRECTIVES: any[] = [container.CONTAINER_DIRECTIVES, filters.FILTER_DIRECTIVES, paging.PAGER_DIRECTIVES];
+export const CARD_CONTAINER_DIRECTIVES: any[] = [container.CONTAINER_DIRECTIVES, filters.FILTER_DIRECTIVES];
 
 export { container, dataSources, filters, paging, sorts };
